Rename: copyZzal 함수, downloadZzal 함수 zzalUtils에 통합
